Improve aiming detection by nearest target. Improved destructible detection by circle area and fov. Melee is canceled when enemy is detected.

+++ Assets/Scripts/UI/Gameplay/UI_PlayerHealthPanel.cs
@@ -64,7 +64,7 @@ namespace LittleLooters.Gameplay.UI
 		{
 			var hp = Mathf.CeilToInt(_playerHealth.Health);
 			var maxHp = _playerHealth.MaxHealth;
-			var progress = hp / maxHp;
+			var progress = (float)hp / (float)maxHp;
 
 			_slider.value = progress;
